requirements.txt
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
django-cors-headers==3.13.0
drf-spectacular==0.25.1
